+++ apps/example/src/main/index.ts
@@ -3,7 +3,7 @@ import { join } from 'path'
 import { electronApp, optimizer, is } from '@electron-toolkit/utils'
 import icon from '../../resources/icon.png?asset'
 
-import { attachToWindow, setup } from 'pretty-titlebar/config'
+import { attachToWindow, setup } from 'electron-pretty-titlebar/config'
 
 setup() // <- Add this line