+++ apps/example/CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Patch Changes
 
 - Updated dependencies [e8ac5f2]
-  - pretty-titlebar@1.0.0
+  - electron-pretty-titlebar@1.0.0